Bug 2024007 - promote float values in FFmpegVideoDecoder decode stats r=media-playback-reviewers,alwu
Promote mAverageFrameDecodeTime and mAverageFrameDuration members to double
to maintain precision.

Differential Revision: https://phabricator.services.mozilla.com/D288508

UltraBlame original commit: cb6090119bdb42aa2eaed5f6151469923d2feeed

+++ dom/media/platforms/ffmpeg/FFmpegVideoDecoder.h
@@ -2,8 +2,6 @@
 
 
 
-
-
 #ifndef FFmpegVideoDecoder_h_
 #define FFmpegVideoDecoder_h_
 
@@ -260,8 +258,8 @@ class FFmpegVideoDecoder<LIBAV_VER>
    private:
     uint32_t mDecodedFrames = 0;
 
-    float mAverageFrameDecodeTime = 0;
-    float mAverageFrameDuration = 0;
+    double mAverageFrameDecodeTime = 0;
+    double mAverageFrameDuration = 0;
 
     
     const uint32_t mMaxLateDecodedFrames = 15;
